chore: move faucet logic to inside handler
