+++ src/lib.rs
@@ -10,7 +10,6 @@ pub mod constants;
 pub mod diagnostics;
 pub mod error;
 pub mod estimation;
-pub mod faucet;
 pub mod interop;
 pub mod liquidity;
 pub mod metrics;
