+++ src/rpc/relay.rs
@@ -9,7 +9,6 @@ use crate::{
     estimation::{
         build_delegation_override, build_simulation_overrides, fees::approx_intrinsic_cost,
     },
-    faucet::FaucetService,
     provider::ProviderExt,
     signers::Eip712PayLoadSigner,
     transactions::interop::InteropBundle,
@@ -36,11 +35,13 @@ use alloy::{
         eip1559::Eip1559Estimation,
         eip7702::{SignedAuthorization, constants::EIP7702_DELEGATION_DESIGNATOR},
     },
-    primitives::{Address, B256, BlockNumber, Bytes, ChainId, U64, U256, aliases::B192, bytes},
+    primitives::{
+        Address, B256, BlockNumber, Bytes, ChainId, TxKind, U64, U256, aliases::B192, bytes,
+    },
     providers::{DynProvider, Provider, utils::EIP1559_FEE_ESTIMATION_PAST_BLOCKS},
     rlp::Encodable,
     rpc::types::{
-        Authorization,
+        Authorization, TransactionRequest,
         state::{AccountOverride, StateOverridesBuilder},
     },
     sol_types::{SolCall, SolValue},
@@ -54,7 +55,7 @@ use jsonrpsee::{
 use opentelemetry::trace::SpanKind;
 use std::{collections::HashMap, iter, sync::Arc, time::SystemTime};
 use tokio::try_join;
-use tracing::{Instrument, Level, debug, error, instrument, span, warn};
+use tracing::{Instrument, Level, debug, error, info, instrument, span, warn};
 
 use crate::{
     chains::{Chain, Chains},
@@ -168,14 +169,12 @@ impl Relay {
         asset_info: AssetInfoServiceHandle,
         escrow_refund_threshold: u64,
     ) -> Self {
-        let faucet_service = FaucetService::new(chains.clone());
         let inner = RelayInner {
             contracts,
             chains,
             fee_recipient,
             quote_signer,
             funder_signer,
-            faucet_service,
             quote_config,
             price_oracle,
             storage,
@@ -2179,11 +2178,86 @@ impl RelayApiServer for Relay {
         &self,
         parameters: AddFaucetFundsParameters,
     ) -> RpcResult<AddFaucetFundsResponse> {
-        self.inner
-            .faucet_service
-            .add_faucet_funds(parameters)
+        let AddFaucetFundsParameters { token_address, address, chain_id, value } = parameters;
+
+        info!(
+            "Processing faucet request for {} on chain {} with amount {}",
+            address, chain_id, value
+        );
+
+        let chain =
+            self.inner.chains.get(chain_id).ok_or(RelayError::UnsupportedChain(chain_id))?;
+
+        // Disallow faucet usage on mainnet chains
+        if alloy_chains::Chain::from(chain_id).named().is_some_and(|c| !c.is_testnet()) {
+            warn!("Faucet request blocked on mainnet (chain {chain_id})");
+            return Ok(AddFaucetFundsResponse {
+                transaction_hash: None,
+                message: Some("Faucet disabled on mainnet".to_string()),
+            });
+        }
+
+        // Token must be a configured fee token on this chain
+        let fee_tokens = chain.assets().fee_tokens();
+        if !fee_tokens.iter().any(|(_, d)| d.address == token_address) {
+            error!("Token address {} not supported for chain {}", token_address, chain_id);
+            return Ok(AddFaucetFundsResponse {
+                transaction_hash: None,
+                message: Some("Token address not supported".to_string()),
+            });
+        }
+
+        // Build calldata for mint(recipient, value)
+        let calldata: Bytes = IERC20::mintCall { recipient: address, value }.abi_encode().into();
+
+        // Estimate gas; if it fails, treat as not supported (e.g., token lacks mint or requires
+        // role)
+        let gas_limit = match chain
+            .provider()
+            .estimate_gas(
+                TransactionRequest::default().to(token_address).input(calldata.clone().into()),
+            )
             .await
-            .map_err(|e| RelayError::InternalError(e).into())
+        {
+            Ok(g) => g,
+            Err(e) => {
+                error!(
+                    "Faucet mint not supported for token {token_address} on chain {chain_id}: {e}"
+                );
+                return Ok(AddFaucetFundsResponse {
+                    transaction_hash: None,
+                    message: Some("Token address not supported".to_string()),
+                });
+            }
+        };
+
+        // Build internal transaction; TransactionService will pick an active relay signer
+        let relay_tx = RelayTransaction::new_internal(
+            TxKind::Call(token_address),
+            calldata,
+            chain.id(),
+            gas_limit,
+        );
+
+        // Send and wait for confirmation
+        let handle = chain.transactions().clone();
+        let _ = handle.send_transaction(relay_tx.clone()).await.map_err(RelayError::from)?;
+        let status = handle.wait_for_tx(relay_tx.id).await.map_err(|_| {
+            RelayError::InternalError(eyre::eyre!("failed to wait for transaction"))
+        })?;
+
+        if !status.is_confirmed() {
+            error!("Faucet funding failed");
+            return Ok(AddFaucetFundsResponse {
+                transaction_hash: status.tx_hash(),
+                message: Some("Faucet funding failed".to_string()),
+            });
+        }
+
+        Ok(AddFaucetFundsResponse {
+            transaction_hash: status.tx_hash(),
+            message: Some("Faucet funding successful".to_string()),
+        })
     }
 }
 
@@ -2200,8 +2274,6 @@ pub(super) struct RelayInner {
     quote_signer: DynSigner,
     /// The signer used to sign fund transfers.
     funder_signer: DynSigner,
-    /// The faucet service for distributing test tokens.
-    faucet_service: FaucetService,
     /// Quote related configuration.
     quote_config: QuoteConfig,
     /// Price oracle.
